Progress towards integrating with new AI Schemas
The GuruVideo class will become the primary interface in this SDK once complete. It integrates with the AI Schemas that customers can create from the console, and allows you to run the same code on browser and device. Still todo is the analysis code, and integrating with the model store.