Include models in manifest & add JS code in bundle (#5)
Pass both the name and the absolute path to each ONNX model in the
manifest when we initialize the engine. Also add the bundle that
includes the Javascript source code for the Guru stdlib and the
node_modules. I'm already starting to question whether a .bundle is the
best option vs. downloading a .zip from an S3 bucket, but we'll go with
it for now.